+++ src/components/input-elements/Select/Select.tsx
@@ -1,9 +1,10 @@
 "use client";
 
-import React, { isValidElement, useCallback, useEffect, useMemo, useState } from "react";
+import React, { isValidElement, useCallback, useMemo, useState } from "react";
 import { ArrowDownHeadIcon, XCircleIcon } from "assets";
 import { makeClassName, tremorTwMerge } from "lib";
 import { constructValueToNameMapping, getSelectButtonColors, hasValue } from "../selectUtils";
+import { HiddenInput } from "../selectUtils";
 
 import { Listbox, Transition } from "@headlessui/react";
 import { useInternalState } from "hooks";
@@ -200,57 +201,3 @@ const Select = React.forwardRef<HTMLDivElement, SelectProps>((props, ref) => {
 Select.displayName = "Select";
 
 export default Select;
-
-export const HiddenInput = ({
-  name,
-  required,
-  value,
-  onReset,
-  setInvalid,
-}: {
-  name?: string;
-  required?: boolean;
-  value?: string;
-  onReset: () => void;
-  setInvalid: (invalid: boolean) => void;
-}) => {
-  const [form, setForm] = useState<HTMLFormElement | null>(null);
-  const [input, setInput] = useState<HTMLInputElement | null>(null);
-
-  useEffect(() => {
-    if (!input) return;
-    const onInvalid = (e: Event) => {
-      e.preventDefault();
-      setInvalid(true);
-    };
-    input.addEventListener("invalid", onInvalid);
-    return () => input.removeEventListener("invalid", onInvalid);
-  }, [input, setInvalid]);
-
-  useEffect(() => {
-    if (!form) return;
-    const reset = () => {
-      setInvalid(false);
-      onReset();
-    };
-    form.addEventListener("reset", reset);
-    return () => form.removeEventListener("reset", reset);
-  }, [form, onReset, setInvalid]);
-
-  return (
-    <input
-      className="hidden"
-      type="text"
-      name={name}
-      required={required}
-      defaultValue={value}
-      tabIndex={-1}
-      ref={(el) => {
-        if (el) {
-          setInput(el);
-          setForm(el.closest("form"));
-        }
-      }}
-    />
-  );
-};
